% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{doping}
\alias{doping}
\title{doping}
\format{
Data frame with 20 rows and 4 variables.

\describe{
  \item{survey}{letter denoting the survey}
  \item{Q4}{Responses to the question about having used doping}
  \item{p4}{Probability of answering yes to the number sequence question.}
  \item{n}{the observed frequencies}
}
}
\source{
WADA
}
\usage{
doping
}
\description{
Doping question Q4 of survey A, B, C, D and E..
}
\keyword{datasets}
