% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medicine}
\alias{medicine}
\title{medicine}
\format{
Data frame with 20 rows and 4 variables.

\describe{
  \item{survey}{letter denoting the survey}
  \item{Q3}{Responses to the question about having used doping}
  \item{p3}{Probability of answering yes to the number sequence question.}
  \item{n}{the observed frequencies}
}
}
\source{
WADA
}
\usage{
medicine
}
\description{
Medicine use question Q3 of survey A, B, C, D and E.
}
\keyword{datasets}
