% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_power}
\alias{plot_power}
\title{Power curves for one-sayers model}
\usage{
plot_power(pi, theta, p, n = seq(10, 1000, by = 10), alpha = 0.05)
}
\arguments{
\item{pi}{prevalence \eqn{\pi} of the sensitive attribute.}

\item{theta}{prevalence \eqn{\theta} of one-saying.}

\item{p}{probability of answering the innocuous question with 'yes' in one
sub-sample and with 'no' in the other.}

\item{n}{vector with sample sizes for which the power is to be computed.}

\item{alpha}{significance level,}
}
\description{
Renders a plot with the power curves of \eqn{\hat\pi} and
\eqn{\hat\theta} given the null hypotheses \eqn{H_0:\pi=0} and \eqn{H_0:\theta=0}.
}
\examples{
plot_power(pi = .2, theta = .1, p = .2)
}
