% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples}
\alias{samples}
\title{samples}
\format{
Data frame with 20 rows and 4 variables.

\describe{
  \item{survey}{letter denoting the survey}
  \item{Q1}{Responses to the question about having used doping}
  \item{p1}{Probability of answering yes to the number sequence question.}
  \item{n}{the observed frequencies}
}
}
\source{
WADA
}
\usage{
samples
}
\description{
Blood/urine question Q1 of survey A, B, C, D and E.
}
\keyword{datasets}
