% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sp_one}
\alias{sp_one}
\title{One-sayers model for the ECWM}
\usage{
sp_one(f0, f1 = ~1, data, pvar)
}
\arguments{
\item{f0}{formula in the form \code{response ~ terms}, where
\code{response} is the numeric variable containing the observed responses 1
(one 'yes' answer and one 'no' answer) or 2 (two 'yes' or two 'no' answers),
and \code{terms} its linear predictor.}

\item{f1}{formula in the form \code{~ terms} with the rhs containing
the linear predictor of the probability of 'self-protective one-saying'.
Defaults to the intercept-only model \code{~ 1}.}

\item{data}{data frame with individual records containing the response
variable, the variable \code{pvar}, and the predictors (if any).}

\item{pvar}{variable containing the probabilities of answering
'yes' to the innocuous (number sequence) question in the subsample.}
}
\value{
List with the parameter estimates and g.o.f. statistics of the fitted models.
}
\description{
Fits the ECWM, the one-sayers model and, if predictors are
included in one or both formulas, the
logistic regression version of the one-sayers model.
}
\examples{
## Intercepts-only models for survey A
sp_one(Q4 ~ 1, ~ 1, dplyr::slice(doping[1:4, ], rep(1:4, n)), p4)

## Model for all surveys with survey as predictor of the prevalence and one-saying
sp_one(Q4 ~ survey, ~ survey, dplyr::slice(doping, rep(1:20, n)), p4)
}
